유저 업데이트 API, 삭제 API 개발 및 테스트 - 유저 삭제 업데이트 API 예외처리

+++ src/main/java/com/group/libraryapp/controller/user/UserController.java
@@ -65,6 +65,14 @@ public class UserController {
     //삭제
     @DeleteMapping("/user")
     public void deleteUser(@RequestParam String name) {
+
+        String readSql = "SELECT * FROM user WHERE name = ?";
+        boolean isUserNotExist = jdbcTemplate.query(readSql, (rs, rowNum) -> 0, name).isEmpty();
+
+        if (isUserNotExist) {
+            throw new IllegalArgumentException();
+        }
+
         String sql = "DELETE FROM user WHERE name = ?";
         jdbcTemplate.update(sql, name);
     }
